[libc] Make socklen_t 32 bits wide
POSIX asks only for an integer type at least 32 bits wide, and the
comment took that as leave to use unsigned long. It is not free to
choose. socklen_t is what a program hands the kernel and what the kernel
writes back through, and it is in the interfaces of every other library
the program is linked with; everywhere this runs, that type is 32 bits.

The layouts came out the same by accident, since the padding that
follows a 32 bit member is exactly what the wider one occupied, and a
value the kernel writes back filled only half the object and was read
correctly only because the machine is little-endian and the caller had
zeroed the rest. What did not survive is the type itself: iptables
carries its own

  typedef unsigned int socklen_t;

for platforms without one, and clang rejects it against a definition
that disagrees.

+++ libc/include/llvm-libc-types/socklen_t.h
@@ -9,10 +9,12 @@
 #ifndef LLVM_LIBC_TYPES_SOCKLEN_T_H
 #define LLVM_LIBC_TYPES_SOCKLEN_T_H
 
-// The posix standard only says of socklen_t that it must be an integer type of
-// width of at least 32 bits. The long type is defined as being at least 32
-// bits, so an unsigned long should be fine.
+// POSIX asks only for an integer type at least 32 bits wide, but the width is
+// not free to choose: socklen_t sits in structures the kernel reads and is
+// pointed at by arguments the kernel writes through, and it appears in the
+// interfaces of every other library a program is linked with. Everywhere this
+// runs, that type is 32 bits wide.
 
-typedef unsigned long socklen_t;
+typedef unsigned int socklen_t;
 
 #endif // LLVM_LIBC_TYPES_SOCKLEN_T_H
